Lowercased modules, Renamed ArenaLayout to field, removed commented code.
