+++ data/field/field.mtl
@@ -1,4 +1,4 @@
-# Blender MTL File: 'ArenaLayout.blend'
+# Blender MTL File: 'field.blend'
 # Material Count: 1
 
 newmtl Material.001
@@ -10,4 +10,4 @@ Ke 0.000000 0.000000 0.000000
 Ni 1.000000
 d 1.000000
 illum 2
-map_Kd ArenaLayout.png
+map_Kd field.png
