+++ simulator/utilities.py
